requirements.txt
# Extraction Layer Dependencies

# PDF processing
pdfplumber==0.10.3
pdf2image==1.17.1

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Logging and utilities
python-dotenv==1.0.0
